+++ packages/css-tokenizer/src/consume/escaped-code-point.ts
@@ -2,14 +2,14 @@ import { MAXIMUM_ALLOWED_CODEPOINT, REPLACEMENT_CHARACTER } from '../code-points
 import { isHexDigitCodePoint, isSurrogate, isWhitespace } from '../code-points/ranges';
 import { CodePointReader } from '../interfaces/code-point-reader';
 import { Context } from '../interfaces/context';
-import { ParseError } from '../interfaces/error';
+import { ParseError, ParseErrorMessage } from '../interfaces/error';
 
 // https://www.w3.org/TR/2021/CRD-css-syntax-3-20211224/#consume-escaped-code-point
 export function consumeEscapedCodePoint(ctx: Context, reader: CodePointReader): number {
 	const codePoint = reader.readCodePoint();
 	if (codePoint === false) {
 		ctx.onParseError(new ParseError(
-			'Unexpected EOF while consuming an escaped code point.',
+			ParseErrorMessage.UnexpectedEOFInEscapedCodePoint,
 			reader.representationStart,
 			reader.representationEnd,
 			[