+++ plugin-packs/postcss-preset-env/src/test/lib/list-features/preserve.mjs
@@ -21,7 +21,6 @@ assert.deepStrictEqual(
 			],
 			vendors_implementations: 3,
 			pluginOptions: {
-				subFeatures: { areaHrefNeedsFixing: true },
 				preserve: true,
 				enableProgressiveCustomProperties: false,
 			},
@@ -38,11 +37,14 @@ assert.deepStrictEqual(
 
 testLogger.logger.emitLogs(testLogger);
 assert.deepStrictEqual(
-	testLogger.getLogs(),
+	testLogger.getLogs().map(x => x.split('\n')),
 	[
-		'Using features from Stage 0\n' +
-		'  blank-pseudo-class has been disabled by "enableClientSidePolyfills: false".\n' +
-		'Adding area[href] fallbacks for ":any-link" support in Edge and IE.',
+		[
+			'Using features from Stage 0.',
+			'- \'blank-pseudo-class\' disabled because \'enableClientSidePolyfills\' is \'false\'.',
+			'- \'any-link-pseudo-class\' enabled for:',
+			'    op_mini all',
+		],
 	],
 );
 
@@ -63,7 +65,6 @@ assert.deepStrictEqual(
 			],
 			vendors_implementations: 3,
 			pluginOptions: {
-				subFeatures: { areaHrefNeedsFixing: true },
 				preserve: false,
 				enableProgressiveCustomProperties: false,
 			},
@@ -80,11 +81,14 @@ assert.deepStrictEqual(
 
 testLogger.logger.emitLogs(testLogger);
 assert.deepStrictEqual(
-	testLogger.getLogs(),
+	testLogger.getLogs().map(x => x.split('\n')),
 	[
-		'Using features from Stage 0\n' +
-		'  blank-pseudo-class has been disabled by "enableClientSidePolyfills: false".\n' +
-		'Adding area[href] fallbacks for ":any-link" support in Edge and IE.',
+		[
+			'Using features from Stage 0.',
+			'- \'blank-pseudo-class\' disabled because \'enableClientSidePolyfills\' is \'false\'.',
+			'- \'any-link-pseudo-class\' enabled for:',
+			'    op_mini all',
+		],
 	],
 );