+++ plugin-packs/postcss-bundler/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Bundler
 
+### Unreleased (major)
+
+- Updated: Support for Node v18+ (major).
+
 ### 1.0.14
 
 _June 29, 2024_